backend/requirements.txt
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
attrs==25.3.0
boto3==1.38.46
botocore==1.38.46
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
cohere==5.11.0
colorama==0.4.6
dill==0.4.0
docker==7.1.0
fastapi==0.115.0
fastavro==1.11.1
filelock==3.18.0
fsspec==2025.5.1
google-pasta==0.2.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.1
idna==3.10
importlib-metadata==6.11.0
jmespath==1.0.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
markdown-it-py==3.0.0
mdurl==0.1.2
mock==4.0.3
multiprocess==0.70.18
mysql-connector-python==9.0.0
numpy==1.26.4
omegaconf==2.3.0
packaging==24.2
pandas==2.3.0
parameterized==0.9.0
pathos==0.3.4
platformdirs==4.3.8
pox==0.3.6
ppft==1.7.7
protobuf==5.29.5
psutil==7.0.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
pywin32==310
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rich==14.0.0
rpds-py==0.25.1
s3transfer==0.13.0
sagemaker==2.247.1
sagemaker-core==1.0.40
schema==0.7.7
six==1.17.0
smdebug-rulesconfig==1.0.1
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.38.6
tblib==3.1.0
tokenizers==0.21.2
tqdm==4.67.1
types-requests==2.32.4.20250611
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.30.6
zipp==3.23.0
